refactor: rename namespace 'Users\*' to 'User\*' and corresponding folder structure
