+++ src/task/Task.php
@@ -2,7 +2,7 @@
 
 namespace R794021\Task;
 
-use R794021\Users\{Contractor, Customer, User};
+use R794021\User\{Contractor, Customer, User};
 use R794021\Action\
     {Action, ApplyAction, CancelAction, DoneAction, RejectAction};
