+++ src/tests/users/CustomerTest.php
@@ -1,5 +1,5 @@
 <?php
-use R794021\Users\{Customer, User};
+use R794021\User\{Customer, User};
 
 $customer = new Customer(CUSTOMER_1);
 assert($customer instanceof User);
